// src/domain/tournament/CostAndChip.interface.ts
export interface CostAndChip {
  title: string;
  num: number;
  cost: number;
  chip: number;
}

// const data = [
//     {
//         title: "buy-in",
//         cost: 50000,
//         chip: 60000
//     },
//     {
//         title: "re-buy-in",
//         cost: 50000,
//         chip: 70000
//     },
//     {
//         title: "re-entry",
//         cost: 50000,
//         chip: 70000
//     },
//     {
//         title: "add-on",
//         cost: 30000,
//         chip: 100000
//     },
// ];